+++ app/src/main/java/app/com/cvjuanresume/juansandoval/cvjuanresume/fragments/FriendsFragment.java
@@ -1,4 +1,4 @@
-package app.com.cvjuanresume.juansandoval.cvjuanresume;
+package app.com.cvjuanresume.juansandoval.cvjuanresume.fragments;
 
 import android.app.Activity;
 import android.support.v4.app.Fragment;
@@ -6,6 +6,8 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import app.com.cvjuanresume.juansandoval.cvjuanresume.R;
 
 /**
  * Created by jsandoval on 18/04/17.
