+++ app/src/main/java/app/com/cvjuanresume/juansandoval/cvjuanresume/fragments/HomeFragment.java
@@ -1,19 +1,22 @@
-package app.com.cvjuanresume.juansandoval.cvjuanresume;
-
-import android.app.Activity;
-import android.support.v4.app.Fragment;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
+package app.com.cvjuanresume.juansandoval.cvjuanresume.fragments;
 
 /**
  * Created by jsandoval on 18/04/17.
  */
 
-public class MessageFragment extends Fragment {
+import android.app.Activity;
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
-    public MessageFragment() {
+import app.com.cvjuanresume.juansandoval.cvjuanresume.R;
+
+
+public class HomeFragment extends Fragment {
+
+    public HomeFragment() {
         // Required empty public constructor
     }
 
@@ -26,7 +29,7 @@ public class MessageFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_messages, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_home, container, false);
 
 
         // Inflate the layout for this fragment
