+++ app/src/main/java/app/com/cvjuanresume/juansandoval/cvjuanresume/fragments/MessageFragment.java
@@ -1,20 +1,21 @@
-package app.com.cvjuanresume.juansandoval.cvjuanresume;
+package app.com.cvjuanresume.juansandoval.cvjuanresume.fragments;
+
+import android.app.Activity;
+import android.support.v4.app.Fragment;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import app.com.cvjuanresume.juansandoval.cvjuanresume.R;
 
 /**
  * Created by jsandoval on 18/04/17.
  */
 
-import android.app.Activity;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
+public class MessageFragment extends Fragment {
 
-
-public class HomeFragment extends Fragment {
-
-    public HomeFragment() {
+    public MessageFragment() {
         // Required empty public constructor
     }
 
@@ -27,7 +28,7 @@ public class HomeFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_home, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_messages, container, false);
 
 
         // Inflate the layout for this fragment
